requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
openpyxl
fpdf2
